backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
bcrypt==4.1.2
reportlab==4.0.7
Pillow==10.1.0
python-dateutil==2.8.2
